% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coalsp}
\alias{coalsp}
\title{Spatially located coal geochemical analysis}
\format{
A list containing two SpatialPonitsDataFrames
\describe{
\item{train}{Training Data, locations of coal analysis and auxilary data}
\item{newdata}{New data locations with auxilary data}
...
}
}
\source{
Maxwell (2020)
}
\usage{
coalsp
}
\description{
A dataset containing coal geochemical analysis and associated auxiliary information
coalsp.
}
\keyword{datasets}
