% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gqrf.R
\name{gqrf}
\alias{gqrf}
\title{Geographic quantile regression forest}
\usage{
gqrf(
  formula,
  data,
  newdata,
  k = NULL,
  ntree = 500,
  mtry = NULL,
  quantiles = c(0.05, 0.5, 0.95)
)
}
\arguments{
\item{formula}{Formula in the format y~X where y is the target variable and X are predictor/s}

\item{data}{Training data of class SpatialPointsDataFrame (sp)}

\item{newdata}{Data of class SpatialPointsDataFrame (sp)comprising of new values to be estimated}

\item{k}{Number of nearest neighbors to use. The default is the number of rows in data minus 1}

\item{ntree}{Number of trees.}

\item{mtry}{Number of variables to possibly split at in each node. Default is the (rounded down) square root of the number variables.}

\item{quantiles}{Vector of the upper, lower and middle quantiles. Default is c(0.05,0.5,0.95)}
}
\description{
Predict values at unknown locations using geographic quantile regression forest
}
\details{
Uses geographic quantile regression forest (Maxwell, 2020) for spatially interpolating values at new, unknown locations.
}
\examples{
#example
}
\references{
\itemize{
\item Maxwell, K., Rajabi, M., Esterle, J. (2020). Spatial interpolation of coal geochemical properties using geographic quantile regression forest. Manuscript submitted for publication.
\item Georganos, S., Grippa, T., Niang Gadiaga, A., Linard, C., Lennert, M., Vanhuysse, S., ... & Kalogirou, S. (2019). Geographical random forests: a spatial extension of the random forest algorithm to address spatial heterogeneity in remote sensing and population modelling. Geocarto International, 1-16
}
}
